Make 'include with' different from 'include for'
The include tag currently accepts the 'with' and 'for' qualifiers when a
local variable is passed in. The documentation at
http://www.rubydoc.info/github/Shopify/liquid/Liquid/Include
implies that these different terms cause different behavior, but this
has never been the case. Consequently, when a variable exists in the
parent template with the same name as the partial it is not possible to
pass an array to that liquid partial without the partial being rendered
once for each element in the same-named array.

This change allows the caller to pass an array in using the 'with'
keyword and force the partial to be rendered once. The default behavior
for same-named variables, and the behavior with the for keyword, remain
unchanged.